+++ algosdk/transaction.py
@@ -87,7 +87,9 @@ class Transaction:
         return sig
 
     def estimate_size(self):
-        return len(encoding.msgpack_encode(self)) + constants.num_additional_bytes_after_signing
+        sk, _ = account.generate_account()
+        stx = self.sign(sk)
+        return len(base64.b64decode(encoding.msgpack_encode(stx)))
 
     def dictify(self):
         d = dict()
@@ -209,7 +211,10 @@ class PaymentTxn(Transaction):
                  lease=None, rekey_to=None):
         Transaction.__init__(self,  sender, fee, first, last, note, gen, gh,
                              lease, constants.payment_txn, rekey_to)
-        self.receiver = receiver
+        if receiver:
+            self.receiver = receiver
+        else:
+            raise error.ZeroAddressError
         self.amt = amt
         if (not isinstance(self.amt, int)) or self.amt < 0:
             raise error.WrongAmountType
@@ -226,7 +231,9 @@ class PaymentTxn(Transaction):
             d["amt"] = self.amt
         if self.close_remainder_to:
             d["close"] = encoding.decode_address(self.close_remainder_to)
-        if self.receiver:
+        
+        decoded_receiver = encoding.decode_address(self.receiver)
+        if any(decoded_receiver):
             d["rcv"] = encoding.decode_address(self.receiver)
 
         d.update(super(PaymentTxn, self).dictify())
@@ -728,7 +735,10 @@ class AssetTransferTxn(Transaction):
                  gen=None, flat_fee=False, lease=None, rekey_to=None):
         Transaction.__init__(self,  sender, fee, first, last, note, gen, gh,
                              lease, constants.assettransfer_txn, rekey_to)
-        self.receiver = receiver
+        if receiver:
+            self.receiver = receiver
+        else:
+            raise error.ZeroAddressError
         self.amount = amt
         if (not isinstance(self.amount, int)) or self.amount < 0:
             raise error.WrongAmountType
@@ -748,8 +758,11 @@ class AssetTransferTxn(Transaction):
             d["aamt"] = self.amount
         if self.close_assets_to:
             d["aclose"] = encoding.decode_address(self.close_assets_to)
-        if self.receiver:
+        
+        decoded_receiver = encoding.decode_address(self.receiver)
+        if any(decoded_receiver):
             d["arcv"] = encoding.decode_address(self.receiver)
+
         if self.revocation_target:
             d["asnd"] = encoding.decode_address(self.revocation_target)
 
@@ -1003,8 +1016,7 @@ class Multisig:
             if subsig.signature is not None:
                 verify_key = VerifyKey(subsig.public_key)
                 try:
-                    verify_key.verify(message,
-                                      base64.b64decode(subsig.signature))
+                    verify_key.verify(message, subsig.signature)
                     verified_count += 1
                 except BadSignatureError:
                     return False
@@ -1231,7 +1243,7 @@ class LogicSig:
         else:
             sig, index = LogicSig.single_sig_multisig(self.logic, private_key,
                                                       multisig)
-            multisig.subsigs[index].signature = sig
+            multisig.subsigs[index].signature = base64.b64decode(sig)
             self.msig = multisig
 
     def append_to_multisig(self, private_key):
@@ -1250,7 +1262,7 @@ class LogicSig:
             raise error.InvalidSecretKeyError
         sig, index = LogicSig.single_sig_multisig(self.logic, private_key,
                                                   self.msig)
-        self.msig.subsigs[index].signature = sig
+        self.msig.subsigs[index].signature = base64.b64decode(sig)
 
     def __eq__(self, other):
         if not isinstance(other, (